feat(SuperBuild): White list filters, OpenMP, Python
Set `TTK_SUPERBUILD_BUILD` cmake variable when TTK is built inside the
ParaView superbuild. In that case, install are disabled and each filter
needs to be explicitly enabled. Additionally, the find python is a bit
different inside the superbuild so a fix have been introduced.

The OpenMP handling has also been improved so that TTK disable OpenMP
when it is not able to find it.

Finally, `vtkGetSet.h` needed to be included before using the
`vtkTypeMacro` in the *core/vtk* header files.